+++ CHANGELOG.md
@@ -14,9 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added SFX player to the API.
 - Added default volume option to the settings.
 - Mighty Summoner macro that summons a creature selected by the user.  
+- Heal and remove fatigue for non-owned tokens/actors in the Personal Health Centre.
 ### Fixed
 - Radiation Centre will now inform about Benny use.
 - Small update fix in shape changer that will hopefully update the combat.
+- Combat injuries were not properly applied in soak damage.
 
 ## [0.15.0] - 2022-03-18
 ### Added